+++ commanderbot_ext/lib/json_serializable.py
@@ -1,6 +1,8 @@
 from abc import ABC, abstractmethod
 from typing import Any
 
+__all__ = ("JsonSerializable",)
+
 
 class JsonSerializable(ABC):
     @abstractmethod